% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dePlzMap.R
\name{dePlzMap}
\alias{dePlzMap}
\title{Create a map of Germany or states of Germany that is colored according to values given for
different PLZ regions.}
\usage{
dePlzMap(
  data,
  title = "",
  bundesland = NA,
  bundeslandBorderColor = "gray",
  naVal = NA,
  populationRelative = NA,
  color = "#115e01"
)
}
\arguments{
\item{data}{a vector of PLZs or a data frame containing a column with name "plz" or "PLZ"
and an additional column with a value. If a vector of PLZs is given, the number of occurences
of each PLZ in the vector is used as the value.}

\item{title}{a title for the plot}

\item{bundesland}{a character vector containing the name(s) of German states,
e.g. "Rheinland-Pfalz". If this argument specified, a map of only the specific states is created.
If this argument is not specified, a complete map of Germany is created.}
}
\value{
a plot of Germany or a set of countries in Germany with colored PLZ areas
}
\description{
Create a map of Germany or states of Germany that is colored according to values given for
different PLZ regions.
}
