% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dePlzMap.R
\name{dePlzMap}
\alias{dePlzMap}
\title{Choropleth map of Germany based on PLZ regions}
\usage{
dePlzMap(
  data,
  bundesland = NA,
  bundeslandBorderColor = "gray",
  highColor = "#115e01",
  naVal = NA,
  title = "",
  legendTitle = NA,
  populationRelative = FALSE,
  percentage = populationRelative,
  decimalMark = ",",
  naColor = "grey75",
  naLabel = "No Data"
)
}
\arguments{
\item{data}{A vector of PLZs or a two-column data frame containing a column with name "plz" or "PLZ"
and an additional column with a value. If a vector of PLZs is given, the number of occurences
of each PLZ in the vector is used as the value.}

\item{bundesland}{A character vector containing the name(s) of German states (Bundesländer),
e.g. "Rheinland-Pfalz". If this argument specified, a map of only the specific states is created.
If this argument is not specified, a complete map of Germany is shown.}

\item{bundeslandBorderColor}{A color for the border around the states, default is \code{"gray"}.}

\item{highColor}{The color for the highest value in a plot with continuous values.}

\item{naVal}{If this argument is set to a value that is not \code{NA}, this value is used in place of \code{NA} values.
The value is also used for PLZ regions that are not listed in the input data.}

\item{title}{The main title for the plot.}

\item{legendTitle}{A title for the legend.
By default, the column name of the value column in the input data frame is used.}

\item{populationRelative}{If set to \code{TRUE}, the values given for the PLZ regions are divided by the
local number of inhabitants. The population data currently used is based on the Registerzensus 2011.}

\item{percentage}{If set to \code{TRUE}, the legend shows the values as percentages.
For example, a value of 1 is displayed as 100 \%.
By default, percentage values are shown if the argument \code{populationRelative} is set to \code{TRUE}.}

\item{decimalMark}{The decimal mark for the legend. The default is \code{","} as in German.}

\item{naColor}{The color to use for displaying missing (\code{NA}) values, defaults to \code{"gray75"}.}

\item{naLabel}{The text of the label in the legend for the missing values.}
}
\value{
A ggplot object representing the plot.
}
\description{
This function creates a map of Germany that is
colored according to values given for different PLZ (postal code) regions.
The appearance of the map and the legend can be customized with different
options.

Input values for display on the map can be continuous (numeric) or discrete.
There is also an option to plot the values relative to the population in the
different PLZ regions, using data from the "Registerzensus 2011" of the
Statistisches Bundesamt in Deutschland.

In addition to plotting a complete map of Germany, it is possible to show
only specific German states (Bundesländer) on the map.
}
